Fixing dependecnies after the refactor
