+++ nauta/dataset/builder.py
@@ -1,8 +1,8 @@
 from torch.utils.data import DataLoader
 
-from nauta.preprocessing import get_preprocessing_layer
+from nauta.dataset.preprocessing import get_preprocessing_layer
 
-from nauta.one_stage.dataset import DeeperShipFeature, DeeperShip
+from nauta.dataset.vtuad import VTUADFeature, VTUAD
 
 def create_data_loader(data, batch_size, shuffle=True):
     """Creates a pytorch dataloader from a Dataset.
@@ -28,14 +28,14 @@ def get_split_dataloader(config, split="test", shuffle=False):
     Returns:
         DataLoader : The desired dataloader object.
     """
-    if config["dataset"]["type"] == "deepershipfeature":
+    if config["dataset"]["type"] == "VTUADfeature":
         batch_size = config["hyperparameters"]["batch_size"]
         dataset_path = config["dataset"][f"{split}_root_path"]
         preprocessings = config["dataset"]["preprocess"]
         num_of_classes = config["model"]["num_of_classes"]
 
         # Get the dataset and dataloader.
-        dataset = DeeperShipFeature(
+        dataset = VTUADFeature(
             dataset_path, num_of_classes=num_of_classes, preprocessing=preprocessings
         )
         dataloader = create_data_loader(dataset, batch_size=batch_size, shuffle=shuffle)
@@ -51,7 +51,7 @@ def get_split_dataloader(config, split="test", shuffle=False):
         transformation = get_preprocessing_layer(pre_processing_type, sample_rate)
 
         # Get the dataset and dataloader.
-        dataset = DeeperShip(
+        dataset = VTUAD(
             metadata_path, sample_rate, number_of_samples, transform=transformation
         )
         dataloader = create_data_loader(dataset, batch_size=batch_size, shuffle=shuffle)
@@ -67,7 +67,7 @@ def get_dataset(config):
     Returns:
         DataLoader, DataLoader : The train and the validation dataloaders, respectively.
     """
-    if config["dataset"]["type"] == "deepershipfeature":
+    if config["dataset"]["type"] == "VTUADfeature":
         batch_size = config["hyperparameters"]["batch_size"]
         train_dataset_path = config["dataset"]["train_root_path"]
         validation_dataset_path = config["dataset"]["validation_root_path"]
@@ -75,12 +75,12 @@ def get_dataset(config):
         num_of_classes = config["model"]["num_of_classes"]
 
         # Get the training and validation.
-        train_dataset = DeeperShipFeature(
+        train_dataset = VTUADFeature(
             train_dataset_path, num_of_classes=num_of_classes, preprocessing=preprocessings
         )
         train_dataloader = create_data_loader(train_dataset, batch_size=batch_size)
 
-        validation_dataset = DeeperShipFeature(
+        validation_dataset = VTUADFeature(
             validation_dataset_path, num_of_classes=num_of_classes, preprocessing=preprocessings
         )
         validation_dataloader = create_data_loader(validation_dataset, batch_size=batch_size, shuffle=False)
@@ -97,12 +97,12 @@ def get_dataset(config):
         transformation = get_preprocessing_layer(pre_processing_type, sample_rate)
 
         # Get the training, validation and test dataloaders.
-        train_dataset = DeeperShip(
+        train_dataset = VTUAD(
             train_metadata_path, sample_rate, number_of_samples, transform=transformation
         )
         train_dataloader = create_data_loader(train_dataset, batch_size=batch_size)
 
-        validation_dataset = DeeperShip(
+        validation_dataset = VTUAD(
             validation_metadata_path, sample_rate, number_of_samples, transform=transformation
         )
         validation_dataloader = create_data_loader(validation_dataset, batch_size=batch_size, shuffle=False)
