+++ nauta/model/__init__.py
@@ -1,0 +1,1 @@
+from nauta.model.builder import get_model
