+++ nauta/tools/__init__.py
@@ -1,3 +1,1 @@
-from nauta.tools.checkpoint import CheckpointManager, Checkpoint
-from nauta.tools.train_manager import TrainManager
 from nauta.tools.utils import create_dir
