+++ nauta/trainer/__init__.py
@@ -1,0 +1,2 @@
+from nauta.tools.checkpoint import CheckpointManager, Checkpoint
+from nauta.tools.train_manager import TrainManager
